Add problems back in
